+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yii Files Change Log
 
-## 1.1.0 under development
+## 2.0.0 under development
+
+- Chg #64: Raise the minimum PHP version to 8.0 and allow `RecursiveDirectoryIterator` as argument in `FileHelper::lastModifiedTime()` (@Gerych1984)
+
+## 1.1.0 
 
 - Bug #58: Add missed return value and type for callback of `set_error_handler()` function (vjik)
 - New #59: Add `beforeCopy`, `afterCopy` callbacks for `FileHelper::copyFile()` and `FileHelper::copyDirectory()` (Gerych1984)